Made minor changes

+++ src/main.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI
+from fastapi import FastAPI, HTTPException
 from typing import List
 from fastapi import Depends
 from src.database import SessionLocal, engine
@@ -35,10 +35,10 @@ def read_income(db: Session = Depends(get_db)):
         return []
     return db_income
 
-@app.get("/income/{income_id}",tags=["INCOME"], response_model=schemas.Income)
+@app.get("/income/{income_id}", tags=["INCOME"], response_model=schemas.Income)
 def read_income(income_id: int, db: Session = Depends(get_db)):
-    db_income = db.query(models.Income).filter(models.Income.income_id==income_id).first()
-    if not db_income:
+    db_income = db.query(models.Income).filter(models.Income.income_id == income_id).first()
+    if db_income is None:
         return []
     return db_income
 
@@ -109,6 +109,7 @@ def delete_expense(expense_id:int , db: Session = Depends(get_db)):
     return {"Message":"ID not available"}
 
 
+
 @app.post("/bank_account/", tags=["BANK ACCOUNT"],response_model=schemas.BankAccount)
 def create_bank_account( bank_account: schemas.BankAccountCreate,db: Session=Depends(get_db)):
     db_account = models.BankAccount(**bank_account.dict())
@@ -117,25 +118,38 @@ def create_bank_account( bank_account: schemas.BankAccountCreate,db: Session=Dep
     db.refresh(db_account)
     return db_account
 
-@app.get("/bank_accounts/{account_id}",tags=["BANK ACCOUNT"], response_model=schemas.BankAccount)
-def read_bank_account( account_id: int,db: Session=Depends(get_db)):
-    return db.query(schemas.BankAccount).filter(models.BankAccount.account_id == account_id).first()
+@app.get("/bank_account/",tags=["BANK ACCOUNT"],response_model=List[schemas.BankAccount])
+def read_bank_account(db:Session=Depends(get_db)):
+    db_account=db.query(models.BankAccount).all()
+    if not db_account:
+        return[]
+    return db_account
 
-@app.put("/bank_accounts/{account_id}", tags=["BANK ACCOUNT"],response_model=schemas.BankAccount)
+@app.get("/bank_account/{account_id}",tags=["BANK ACCOUNT"], response_model=schemas.BankAccount)
+def read_bank_account( account_id: int,db: Session=Depends(get_db)):
+    return db.query(models.BankAccount).filter(models.BankAccount.account_id == account_id).first()
+
+@app.put("/bank_account/{account_id}", tags=["BANK ACCOUNT"],response_model=schemas.BankAccount)
 def update_bank_account(account_id: int, bank_account: schemas.BankAccountCreate,db: Session=Depends(get_db)):
     db_account = db.query(models.BankAccount).filter(models.BankAccount.account_id == account_id).first()
-    for key, value in bank_account.dict().items():
-        setattr(db_account, key, value)
+    if not db_account:
+        return{"message":"Bank account not found"}
+    db_account.user_id=bank_account.user_id
+    db_account.balance=bank_account.balance
+    db_account.name_of_bank=bank_account.name_of_bank
+    db_account.account_no=bank_account.account_no
     db.commit()
     db.refresh(db_account)
     return db_account
 
-@app.delete("/bank_accounts/{account_id}",tags=["BANK ACCOUNT"])
+@app.delete("/bank_account/{account_id}",tags=["BANK ACCOUNT"])
 def delete_bank_account(account_id: int,db: Session=Depends(get_db) ):
     db_account = db.query(models.BankAccount).filter(models.BankAccount.account_id == account_id).first()
-    db.delete(db_account)
-    db.commit()
-    return db_account
+    if db_account:
+        db.delete(db_account)
+        db.commit()
+        return {"Message":"Records deleted"}
+    return {"Message":"ID not available"}
 
 @app.post("/categories/",tags=["CATEGORY"], response_model=schemas.Category)
 def create_category(category: schemas.CategoryCreate,db: Session= Depends(get_db)):
@@ -153,6 +167,14 @@ def read_category(db:Session=Depends(get_db)):
         return[]
     return db_category
 
+@app.get("/categories/{category_id}", tags=["CATEGORY"],response_model=schemas.Category)
+def read_expense(category_id: int, db: Session = Depends(get_db)):
+    db_category = db.query(models.Category).filter(models.Category.category_id==category_id).first()
+    if not db_category:
+        return []
+    return db_category
+
+
 @app.put("/categories/{category_id}",tags=["CATEGORY"],response_model=schemas.Category)
 def update(category_id:int,category: schemas.CategoryCreate,db:Session=Depends(get_db)):
     db_category=db.query(models.Category).filter(models.Category.category_id==category_id).first()
@@ -167,9 +189,9 @@ def update(category_id:int,category: schemas.CategoryCreate,db:Session=Depends(g
 @app.delete("/categories/{category_id}",tags=["CATEGORY"])
 def delete_category(category_id:int,db:Session=Depends(get_db)):
     db_category=db.query(models.Category).filter(models.Category.category_id==category_id).first()
-    if not db_category:
-        return{"Catgeory not found"}
-    db.delete(db_category)
-    db.commit()
-    return db_category
+    if db_category:
+        db.delete(db_category)
+        db.commit()
+        return {"Message":"Records deleted"}
+    return {"Message":"ID not available"}
